Support sd_mod being compiled as module
Currently trying to use zolspinup on a system where sd_mod module is
compiled as a module fails with the following error:

bpf: sd_resume_runtime is not found in vmlinux BTF
...
Exception: Failed to load BPF program
b'kfunc__vmlinux__sd_resume_runtime': No such file or directory

This is because when the target function is in a module, the eBPF
program must use MODULE_KFUNC_PROBE instead of KFUNC_PROBE.

This commit adds support for both cases by checking whether sd_mod is
loaded as a module and if yes, using MODULE_KFUNC_PROBE-based eBPF
program.

sd_mod is compiled as a module in popular Linux distributions by
default, such as on Debian Bookworm and Trixie (BLK_DEV_SD=m).